app/src/main/java/com/aranteknoloji/ktroom/db/AppDatabase.kt
package com.aranteknoloji.ktroom.db

import android.arch.persistence.room.Database
import android.arch.persistence.room.Room
import android.arch.persistence.room.RoomDatabase
import android.content.Context

@Database(entities = arrayOf(User::class), version = 1)
abstract class AppDatabase: RoomDatabase() {

    abstract fun userDao(): UserDao

    companion object {
        private var instance: AppDatabase? = null

        fun getInstance(context: Context): AppDatabase? {
            if (instance == null) {
                instance = Room.databaseBuilder(context.applicationContext,
                        AppDatabase::class.java, "userdatabase")
                        .allowMainThreadQueries()
                        .build()
            }
            return instance
        }
    }
}